requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx==0.25.2
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
groq>=0.5.0
